Shuffle arxiv.arxiv logic into arxiv.api
Update some tests to use the new import style
