+++ arxiv/__init__.py
@@ -1,4 +1,3 @@
 """.. include:: ../README.md"""
 # flake8: noqa
 from .arxiv import *
-from .query import *
